requirements.txt
opencv-python
numpy
matplotlib
seaborn
natsort
ultralytics
torch
segment-anything
scipy
scikit-learn
pyyaml